Refactor JMA repository structure for better consistency.
