+++ soil_analysis/domain/repository/jma.py
